Correction for AverageOfInput

+++ week-02/day-01/src/AverageOfInput.java
@@ -4,21 +4,21 @@ public class AverageOfInput {
     public static void main(String[] args) {
         Scanner myScanner = new Scanner(System.in);
         System.out.println ("Please enter a number" );
-        int a = myScanner.nextInt ();
+        double a = myScanner.nextInt ();
 
         System.out.println ("Please enter a number" );
-        int b = myScanner.nextInt ();
+        double b = myScanner.nextInt ();
 
         System.out.println ("Please enter a number" );
-        int c = myScanner.nextInt ();
+        double c = myScanner.nextInt ();
 
         System.out.println ("Please enter a number" );
-        int d = myScanner.nextInt ();
+        double d = myScanner.nextInt ();
 
         System.out.println ("Please enter a number" );
-        int e = myScanner.nextInt ();
+        double e = myScanner.nextInt ();
 
-        int sum = a + b + c + d + e;
+        double sum = a + b + c + d + e;
         double average = sum / 5;
 
         System.out.println ("Sum is "+ sum );
